refactor: combat phase-2 slice 6 — naval scenario tables (#3865)
Migrate all three naval combat test suites to table-driven scenario
loops in colonizethis_combat_test_support with shared navalTwoPlayerGame
builders via TestFixtures.minimalGame. Sync dev merge included.
