+++ packages/colonizethis_combat_test_support/lib/colonizethis_combat_test_support.dart
@@ -15,3 +15,7 @@ export 'src/quick_battle_input_test_support.dart';
 export 'src/quick_battle_resolver_scenarios.dart';
 export 'src/quick_battle_siege_pipeline_test_support.dart';
 export 'src/quick_battle_siege_scenarios.dart';
+export 'src/naval_combat_test_support.dart';
+export 'src/naval_combat_resolver_scenarios.dart';
+export 'src/naval_combat_resolution_scenarios.dart';
+export 'src/naval_combat_privateering_scenarios.dart';
