+++ packages/colonizethis_combat_test_support/pubspec.yaml
@@ -10,6 +10,7 @@ environment:
 
 dependencies:
   colonizethis_combat:
+  colonizethis_data:
   colonizethis_logic:
   colonizethis_models:
   colonizethis_test:
